Truncate file names
I noticed that if you do a very long query the file names that get generated can be larger than the host permits. So I added some truncation that should work pretty universally to limit the file name to under <250 characters total including the important naming convention of the prefixes.